private key removed from Gen2 struct

+++ bip32/src/address_type.rs
@@ -14,7 +14,7 @@ impl ToString for AddressType {
 }
 
 impl AddressType {
-    pub fn index(&self) -> i8 {
+    pub fn index(&self) -> u32 {
         match self {
             Self::Receive => 0,
             Self::Change => 1,
